requirements.txt
django_csp==3.8
Django==5.1
mysqlclient==2.2.4
pillow==10.4.0
python-dotenv
sqlparse==0.5.1
asgiref==3.8.1
django-crispy-forms==2.2
gunicorn==20.1.0
numpy==2.0.0
pandas==2.2.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-magic==0.4.27
pytz==2024.1
setuptools==70.1.0
six==1.16.0
somepackage==1.2.3
tzdata==2024.1
whitenoise==6.7.0